Add initial code for POSIX-compatible OSes
Drafted the initial SerialPort interface, as well as the specific
SerialPortPOSIX implementation. Each implementation has a
`GenericReader` and a `GenericWriter` for convenience.

This implementation relies on the `termios` header found in libc.
Initialization flags are set according to the description of the
`cfmakeraw` function. Reader and writer use POSIX read/write calls
respectively.